Update Lily adapters and middleware to only expect `Callable`s
This commit breaks from the previous constraints of requiring objects
that implement a ::handler() method and ::wrapHandler() for middleware.
Instead all we require is that you pass `Callable`s around. This might
be a plain anonymous function or it might be an object that implements
`::__invoke()` a PHP method that's been around since PHP 5.3.