Remove normalisation of response logic from Lily\Adapter\HTTP
Normalising response logic in Lily\Adapter\HTTP creates problems when
it comes to middleware. If a handler doesn't always return an
associative array, we get problems adapting the response as we have to
check if it is a string, numeric array or associative array. For now
there will no longer be short forms of responses provided by this class.